backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
yfinance==0.2.28
beautifulsoup4==4.12.2
requests==2.31.0
lxml。。。==4.9.3
flask-mail==0.9.1
PyJWT==2.8.0
nsepython==1.2.21
nsetools==1.0.11
bsedata==1.0.3
